requirements.txt
ace_tools==0.0
acres==0.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Authlib==1.3.1
azure-core==1.32.0
azure-identity==1.19.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bitarray==3.0.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
bm25s==0.2.7.post1
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.20.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docarray==0.40.0
duckduckgo_search==7.5.0
durationpy==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.9.0.post1
fastapi==0.115.7
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.1.24
fonttools==4.55.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.12.0
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.28
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keybert==0.9.0
kiwisolver==1.4.7
kubernetes==32.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.40
langchain-google-genai==2.0.10
langchain-groq==0.2.1
langchain-neo4j==0.3.0
langchain-ollama==0.2.3
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.2
language_data==1.3.0
llama-cloud==0.1.12
llama-cloud-services==0.6.0
llama-index==0.12.16
llama-index-agent-openai==0.4.5
llama-index-cli==0.4.0
llama-index-core==0.12.16.post1
llama-index-embeddings-azure-openai==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-azure-openai==0.3.0
llama-index-llms-openai==0.3.18
llama-index-multi-modal-llms-openai==0.4.3
llama-index-postprocessor-colbert-rerank==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.2
llama-index-vector-stores-qdrant==0.6.0
llama-parse==0.6.0
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.11
mistune==3.0.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
mypy-extensions==1.0.0
narwhals==1.14.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nipype==1.9.2
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.20.1
openai==1.60.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
posthog==3.10.0
preshed==3.0.9
primp==0.14.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.0
prov==2.0.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.3
Pygments==2.18.0
PyJWT==2.10.1
pymilvus==2.5.4
PyMuPDF==1.25.1
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
PyStemmer==2.2.0.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.13.3
ragas==0.2.13
rank-bm25==0.2.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
segtok==1.5.11
selenium==4.29.0
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.5.0
stack-data==0.6.3
starlette==0.45.3
streamlit==1.40.2
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
toolz==1.0.0
torch==2.2.0+cpu
torchvision==0.17.0+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
transformers==4.46.3
trio==0.29.0
trio-websocket==0.12.1
triton==3.1.0
typer==0.15.1
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==4.10.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.14.2
xxhash==3.5.0
yake==0.4.8
yarl==1.18.0
zipp==3.21.0
zstandard==0.23.0
